refactor: migrate SwapSlippage components to the kit package
- Removed `SwapSlippage` components from the components package and added them to the kit package for better organization.
- Updated import paths across the codebase to reflect the new locations of `SlippageInput`, `SlippageSettingDialog`, and the validation utility.
- Ensured that the functionality remains intact while streamlining the component structure.
