+++ packages/kit/src/utils/validateAmountInput.ts
