+++ packages/kit/src/views/Staking/components/UniversalWithdraw/index.tsx
@@ -26,12 +26,12 @@ import {
   calcPercentBalance,
 } from '@onekeyhq/kit/src/components/PercentageStageOnKeyboard';
 import { usePromiseResult } from '@onekeyhq/kit/src/hooks/usePromiseResult';
+import { validateAmountInput } from '@onekeyhq/kit/src/utils/validateAmountInput';
 import { useSettingsPersistAtom } from '@onekeyhq/kit-bg/src/states/jotai/atoms';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
 import earnUtils from '@onekeyhq/shared/src/utils/earnUtils';
 import type { IEarnEstimateFeeResp } from '@onekeyhq/shared/types/staking';
 
-import { validateAmountInput } from '../../../Swap/utils/utils';
 import { capitalizeString, countDecimalPlaces } from '../../utils/utils';
 import { EstimateNetworkFee } from '../EstimateNetworkFee';
 import { StakingAmountInput } from '../StakingAmountInput';
