+++ packages/kit/src/views/Swap/pages/components/SwapHeaderRightActionContainer.tsx
@@ -24,7 +24,7 @@ import {
   HeaderButtonGroup,
   HeaderIconButton,
 } from '@onekeyhq/components/src/layouts/Navigation/Header';
-import { SlippageInput } from '@onekeyhq/components/src/SwapSlippage';
+import { SlippageInput } from '@onekeyhq/kit/src/components/SwapSlippage';
 import useAppNavigation from '@onekeyhq/kit/src/hooks/useAppNavigation';
 import { useSwapTypeSwitchAtom } from '@onekeyhq/kit/src/states/jotai/contexts/swap';
 import {
